Added Skills version support on /v1/skills + added OpenAI shell support + added Azure (blob storage) Skill provider + updated Perplexity models + added AI21 Maestro + added RaxAI, AIDuet, PiAPI, YouCom, IApp, OODAAI, StealthGPT, Messari, JassieAI, MemoryRouter, ImageRouter, Cailos, CodexForMe, RunAPI, BaseAPI, YYClaw, GateRouter, TeamDay, OXOAPI, Chutes, XiaomiMIMO, Ollama, SwitchpointAI, Vivgrid, Ambient, Relace, Inflection, LangDB, Baidu, Morph, StreamLake, AkashML backend providers + updated to Vercel V7 specs + improved model listing refresh + various other fixes

+++ Core/AIHappey.Core/Providers/LaoZhang/LaoZhangProvider.Models.cs
@@ -54,7 +54,7 @@ public partial class LaoZhangProvider
                     }
 
                     if (el.TryGetProperty("created", out var createdEl))
-                        model.Created = createdEl.GetInt32();
+                        model.Created = createdEl.GetInt64();
 
                     if (el.TryGetProperty("owned_by", out var orgEl))
                         model.OwnedBy = orgEl.GetString() ?? "";
